requirements.txt
geopandas==0.12.1
pandas==1.5.1
python-Levenshtein==0.20.8
thefuzz==0.19.0
tqdm==4.64.1
usaddress==0.5.10
usaddress-scourgify==0.2.4
